feat(types): swap order of Input and Def in ZodType
BREAKING CHANGE:

I recently raised an issue (#397) where I learned that I needed to
specify the `Input` type parameter for `Schema`/`ZodType` in order to
prevent it being inferred as a union of the input and ouput types of a
call to `.transform()`.

I missed figuring this out for myself as it required me understanding
and providing the `Def` parameter also as it comes before the `Input`
type in the type argument list.

On inspection it feels like it is important for `Def` to be inferred
correctly as it changes the interface of the schema type and therefore
as a user I probably shouldn't try and specify it. However, it does feel
like specifying the `Input` type is a reasonable thing to want to do.

This patch swaps the order of the `Def` and `Input` types in `ZodType`
so that `Input` can be specified whilst leaving `Def` to be inferred by
the compiler.

This would require specifying the same type twice in cases where you
would want to specify the `Def` type, so this also adds a
`ZodTypeSymmetric<T, Def = ZodTypeDef>` that sets the `Input` and
`Output` type parameters to `T`, to prevent this duplication.

I appreciate that is a a breaking change, but I believe it is a
worthwile quality-of-life improvement that would great to get this in
before 3.0 is properly released.