Return invalid_union on dirty results
When one or more of the results in the union is dirty the union would
previously just return that result. However this can make the error
message strange and inconvenient to override in some cases, drops the
rest of the dirty results if there are multiple, and hides the fact that
it's a union.

Therefore, it's changed to return an invalid_union issue with all the
dirty results. When there are some dirty results, the aborted results
are not included, since they are not relevant. When there are no dirty
results, all the results are included as before.

An example of an error message that's strange and inconvenient to
override is if you have a union of two strict objects, e.g.:

    z.union([
        z.object({ a: z.number() }).strict(),
        z.object({ b: z.number() }).strict()
    ])

If the input is `{ a: 1, b: 1 }`, it would previously return the result
from the first object, which would be code unrecognized_keys with keys
`b`. If you wanted to override the error message, you would have to do
that in the first object, it wouldn't suffice to override the message in
union.

With this patch, the result would be code invalid_union with the two
unrecognized_keys in the unionErrors array. Since the error is now
always invalid_union it suffices to override the error message in union
for both this case when both keys are provided, and for the case when
none of the union schemas match.

Fixes #866