Allow wrapped literal types in discriminatedUnion
This fixes type checking when passing literal types with default values or preprocess transformations.

Before this change, the type checker was complaining that value does not exist, as described in https://github.com/colinhacks/zod/issues/1490.

The fix consists in modifying the `ZodDiscriminatedUnionOption` to also allow `ZodDefault<ZodLiteral...>` and `ZodEffects<ZodLiteral...>`.

Added a test case that had failing types but now works.

Note: An even better solution I imagine is making sure that both `ZodDefault` and `ZodEffects` re-surface the inner type properties, such as `ZodLiteral`'s `value`. But I'd need more familiarity with the code base to propose that change.