Fix branded Record keys in ZodRecord
Because of a quirk of the extends syntax of typescript conditional
types, the previous version of this check for a branded key didn't work
as expected.

Thank you @akomm for the explanation of why the extends sides need to be
swapped at https://github.com/colinhacks/zod/pull/2287#issuecomment-1689922672

Originally I intended to remove the RecordType type entirely, since it
would seem initially that it is only there to work around earlier
versions of typescript not having the `noUncheckedIndexedAccess` option.
However, it is really meant to work around a problem with using ZodEnum
as the key, see this comment:
https://github.com/colinhacks/zod/pull/2287#issuecomment-1676624932