Add `zocker` to README
About a week ago I posted on the zod-discord about my project Zocker.
It was then suggested I make a PR to add it to the added to the Ecosystem section, so here I am. 


Zocker is a mock-data generation library that generates plausible mock-data from zod-schemas, kinda like `@anatine/zod-mock`, but supports a wider subset of zod.