Wrap preprocess with try-catch to provide better information on error
Before this fix, any errors during preprocess would result in an error
that doesn't provide a lot of debug information. For example, the path
where the error happened was not provided by Zod.
This change will make sure that even errors during preprocess are
translated to ZodError with useful metadata about the error.

Solves issue #2724