Fix globbing in `prettier:*` scripts
The globs in the `prettier:*` scripts in `package.json` need to be
quoted to ensure that the globs are evaluated by Prettier, rather than
the shell. Before, the globs only matched files in _subdirectories_ of
`src/` and `deno/lib/`, and not top-level files in those directories.
This is demonstrated by `deno/lib/types.ts` having incorrect formatting.

Also fix globbing errors in the `lint-staged` config and include the
deno files there as well for consistency.