fix(transform): abort further processing when inner value is invalid
Previously when an inner value was invalid, transforms would not run, but
subsequent refinements would still run with the non-transformed value, which
in most cases would violate the typings.

This fix returns `INVALID` from the transform in case of a failed inner
validation and stop outer refinements from running with the wrong types.

Closes #2243
Closes #2192
Closes #2113