Update docs z.enum with object literal example
From the tests, it looks like `z.enum` supports object literal comprehension out-of-the-box:

https://github.com/colinhacks/zod/blob/3048d14bc7b803507302b8ee3cd1eeeffbd27396/packages/zod/src/v4/classic/tests/enum.test.ts#L12-L24

But it doesn't look like the rest of the world realizes it. Googling for "zod z.enum object literal" leads me to stackoverflow discussions ([1](https://stackoverflow.com/questions/73825273/creating-a-zod-enum-from-an-object), [2](https://stackoverflow.com/questions/75217817/zod-enum-and-object-as-const)) that keep trying to use `Object.values` or other stuff to get this working. I don't really understand.

Updates the documentation with an object literal example, in the hopes that this knowledge propagates and people stop trying to hack around something that works out-of-the-box natively.

Note that [valibot's enum docs](https://valibot.dev/guides/enums/) call this out explicitly:

> Since TypeScript enums are transpiled to JavaScript objects by the TypeScript compiler, you can use the `enum` schema function for both. Just pass your enumerated data type as the first argument to the schema function.