Convert `Date` instances to numbers in `minimum`/`maximum` (#5351)
* Add `Date` to `minimum` and `maximum` in too small/big issues

* Convert dates to numbers in `too_big` and `too_small` issues

This change aligns the runtime values with typings. Detection of date
objects is implemented by checking if `typeof` returns `object`, as the
same check is used to initialize `origin`.

Furthermore, this changes initialization of test values so that they do
not rely on local time zone. (`Date` constructor with multiple numeric
parameters interprets them in the local time zone.)